Split SGFX core from platform backends
