+++ user/lib/sgfx-backend-wgpu/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "sgfx-backend-wgpu"
+version = "0.16.0"
+edition = "2024"
+
+[dependencies]
+bytemuck = { version = "1", features = ["derive"] }
+sgfx-core = { path = "../sgfx", default-features = false }
+wgpu = "24"
+
+[dev-dependencies]
+pollster = "0.4"
+
+[lib]
+name = "sgfx_backend_wgpu"
+path = "src/lib.rs"
