+++ user/std-bin/Cargo.toml
@@ -12,7 +12,7 @@ mp4-aac = ["dep:symphonia-core", "dep:symphonia-codec-aac"]
 clap = { version = "4.5.32", features = ["derive"] }
 csv = "1.3"
 framebuffer = { path = "../lib/framebuffer", default-features = false, features = ["std"] }
-sgfx = { path = "../lib/sgfx", default-features = false, features = ["std"] }
+sgfx = { path = "../lib/sgfx-compat", default-features = false, features = ["std"] }
 httparse = "1.9"
 libm = "0.2"
 lyon_path = "1.0"
